Split out/remove CSVMerge & WolvenKit. Stop creating default button.
