+++ src/index.ts
@@ -21,17 +21,6 @@ const moddingTools = [
     shell: false,
     relative: true,
   },
-  {
-    id: "CSVMerge",
-    name: "CSVMerge",
-    executable: () => path.join("csvmerge", "CSVMerge.cmd"),
-    requiredFiles: [
-      path.join("csvmerge", "CSVMerge.cmd"),
-      path.join("csvmerge", "wolvenkitcli", "WolvenKit.CLI.exe"),
-    ],
-    shell: true,
-    relative: true,
-  },
 ];
 
 export const findGame = () =>
