+++ src/installer.core.ts
@@ -9,7 +9,6 @@ import {
 } from "./vortex-wrapper";
 import { instructionsForSameSourceAndDestPaths } from "./installers.shared";
 import { FileTree } from "./filetree";
-import { wolvenKitDesktopFoundErrorDialog } from "./ui.dialogs";
 
 const path = win32;
 
@@ -26,10 +25,6 @@ const RED4EXT_CORE_IDENTIFIERS = [
   path.normalize("red4ext/LICENSE.txt"),
   path.normalize("red4ext/RED4ext.dll"),
 ];
-
-const CSVMERGE_UNIQUE_FILE = path.normalize("csvmerge/CSVMerge.cmd");
-
-const WOLVENKIT_UNIQUE_FILE = path.normalize("WolvenKit CLI/WolvenKit.CLI.exe");
 
 export const testForCetCore: VortexWrappedTestSupportedFunc = (
   api: VortexApi,
@@ -119,91 +114,6 @@ export const installRed4ExtCore: VortexWrappedInstallFunc = (
     destination: path.normalize("red4ext/plugins"),
   });
   const instructions = [].concat(red4extInstructions, pluginsDir);
-
-  return Promise.resolve({ instructions });
-};
-
-export const testCoreCsvMerge: VortexWrappedTestSupportedFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  if (!files.includes(CSVMERGE_UNIQUE_FILE)) {
-    return Promise.resolve({
-      supported: false,
-      requiredFiles: [],
-    });
-  }
-
-  return Promise.resolve({
-    supported: true,
-    requiredFiles: [],
-  });
-};
-
-export const installCoreCsvMerge: VortexWrappedInstallFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-  _destinationPath: string,
-): Promise<VortexInstallResult> => {
-  log("info", "Using CSV installer");
-
-  const instructions = instructionsForSameSourceAndDestPaths(files);
-
-  return Promise.resolve({ instructions });
-};
-
-export const testCoreWolvenKitCli: VortexWrappedTestSupportedFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-): Promise<VortexTestResult> => {
-  if (files.some((file: string) => file.toLowerCase().startsWith("wolvenkit desktop"))) {
-    const message = "WolvenKit Desktop is not able to be installed with Vortex.";
-    wolvenKitDesktopFoundErrorDialog(api, log, message, files, []);
-    return Promise.reject(new Error(message));
-  }
-
-  if (!files.includes(WOLVENKIT_UNIQUE_FILE)) {
-    return Promise.resolve({
-      supported: false,
-      requiredFiles: [],
-    });
-  }
-
-  return Promise.resolve({
-    supported: true,
-    requiredFiles: [],
-  });
-};
-
-export const installCoreWolvenkit: VortexWrappedInstallFunc = (
-  api: VortexApi,
-  log: VortexLogFunc,
-  files: string[],
-  _fileTree: FileTree,
-  _destinationPath: string,
-): Promise<VortexInstallResult> => {
-  log("info", "Using Wolvenkit CLI installer");
-
-  const allWolvenKitFiles = files.filter((file: string) => !file.endsWith(path.sep));
-
-  const wolvenKitInstructions = allWolvenKitFiles.map((file: string) => {
-    const regex = /^WolvenKit CLI/;
-    const dest = file.replace(regex, path.normalize("csvmerge/wolvenkitcli"));
-
-    return {
-      type: "copy",
-      source: file,
-      destination: dest,
-    };
-  });
-
-  const instructions = [].concat(wolvenKitInstructions);
 
   return Promise.resolve({ instructions });
 };
